feat(appeals): add expedited has fields to the PDF generator (A2-8462)
