+++ appeals/pdf-service-api/src/mappers/appellant-case/__tests__/appellant-case.mapper.test.js
@@ -5,6 +5,16 @@ describe('mapAppellantCaseData', () => {
 	describe('appellantCase', () => {
 		it.each([
 			['householder', mockAppellantCaseData.appellantCaseDataHouseholder],
+			[
+				'householder submitted after cutoff',
+				{
+					...mockAppellantCaseData.appellantCaseDataHouseholder,
+					applicationDate: '2026-04-01T00:00:00.000Z',
+					reasonForAppealAppellant: 'My reason for appeal',
+					anySignificantChanges: true,
+					anySignificantChanges_otherSignificantChanges: 'Other changes'
+				}
+			],
 			['s78', mockAppellantCaseData.appellantCaseDataS78],
 			['s78 expedited', mockAppellantCaseData.appellantCaseDataS78Expedited],
 			['s20', mockAppellantCaseData.appellantCaseDataS20],
