+++ appeals/pdf-service-api/src/mappers/appellant-case/sections/application-details/row-keys.js
@@ -1,10 +1,16 @@
 import { APPEAL_TYPE } from '@pins/appeals/constants/common.js';
+import { beforeExpeditedOriginalApplicationCutOff } from '@pins/appeals/utils/appeal-type-checks.js';
+
 export const rowKeys = {
 	[APPEAL_TYPE.HOUSEHOLDER]: [
 		'applicationDate',
 		'developmentDescription',
 		'otherAppealsList',
-		'applicationDecisionLetter'
+		'applicationDecisionLetter',
+		{
+			key: 'reasonForAppealAppellant',
+			condition: (data) => !beforeExpeditedOriginalApplicationCutOff(data.applicationDate)
+		}
 	],
 	[APPEAL_TYPE.CAS_PLANNING]: [
 		'applicationDate',
